+++ libs/components/src/lib/RangePicker/index.spec.tsx
@@ -1,6 +1,6 @@
 import { render } from "@testing-library/react";
 
-import RangePicker from ".";
+import { RangePicker } from ".";
 import "@testing-library/jest-dom/extend-expect";
 
 describe("RangePicker Component", () => {
